+++ dsptoolbox/_standard.py
@@ -1,6 +1,7 @@
 """
 Backend for standard functions
 """
+
 import numpy as np
 from scipy.signal import correlate, check_COLA, windows, hilbert
 from scipy.special import iv as bessel_first_mod
@@ -284,8 +285,6 @@ def _welch(
         if not autospectrum:
             y_frames -= np.mean(y_frames, axis=0)
 
-    # Combine
-
     if not autospectrum:
         sp_frames = np.fft.rfft(x_frames, axis=0).conjugate() * np.fft.rfft(
             y_frames, axis=0
@@ -293,7 +292,7 @@ def _welch(
     else:
         sp_frames = np.abs(np.fft.rfft(x_frames, axis=0)) ** 2
 
-    # Direct averaging much faster
+    # Direct averaging (much faster than averaging magnitude and phase)
     if average == "mean":
         csd = np.mean(sp_frames, axis=1)
     else:
@@ -319,13 +318,10 @@ def _welch(
     else:
         factor = 1
 
-    # Zero frequency fix when detrending (especially useful for dB plotting)
-    if detrend:
-        csd[0, ...] = csd[1, ...]
-
     csd *= factor
-    csd[0, ...] /= 2
-    csd[-1, ...] /= 2
+    if factor != 1:
+        csd[0, ...] /= 2
+        csd[-1, ...] /= 2
 
     if "amplitude" in scaling:
         csd = np.sqrt(csd)
@@ -361,7 +357,10 @@ def _group_delay_direct(phase: np.ndarray, delta_f: float = 1):
 
 
 def _minimum_phase(
-    magnitude: np.ndarray, unwrapped: bool = True
+    magnitude: np.ndarray,
+    whole_spectrum: bool = False,
+    unwrapped: bool = True,
+    odd_length: bool = False,
 ) -> np.ndarray:
     """Computes minimum phase system from magnitude spectrum.
 
@@ -370,8 +369,15 @@ def _minimum_phase(
     magnitude : `np.ndarray`
         Spectrum for which to compute the minimum phase. If real, it is assumed
         to be already the magnitude.
+    whole_spectrum : bool, optional
+        When `True`, it is assumed that the spectrum is passed with both
+        positive and negative frequencies. Otherwise, the negative frequencies
+        are obtained by mirroring the spectrum. Default: `False`.
     uwrapped : bool, optional
         If `True`, the unwrapped phase is given. Default: `True`.
+    odd_length : bool, optional
+        When `True`, it is assumed that the underlying time data of the half
+        spectrum had an odd length. Default: `False`.
 
     Returns
     -------
@@ -381,9 +387,23 @@ def _minimum_phase(
     """
     if np.iscomplexobj(magnitude):
         magnitude = np.abs(magnitude)
+
+    original_length = magnitude.shape[0]
+    if not whole_spectrum:
+        if odd_length:
+            # Nyquist is not contained in the spectrum
+            magnitude = np.concatenate(
+                [magnitude, magnitude[1:][::-1]], axis=0
+            )
+        else:
+            magnitude = np.concatenate(
+                [magnitude, magnitude[1:-1][::-1]], axis=0
+            )
+
     minimum_phase = -np.imag(
         hilbert(np.log(np.clip(magnitude, a_min=1e-40, a_max=None)), axis=0)
-    )
+    )[:original_length]
+
     if not unwrapped:
         minimum_phase = np.angle(np.exp(1j * minimum_phase))
     return minimum_phase
@@ -746,13 +766,46 @@ def _kaiser_window_beta(A):
     """
     A = np.abs(A)
     if A > 50:
-        beta = 0.1102 * (A - 8.7)
-    elif A >= 21:
-        beta = 0.5842 * (A - 21) ** 0.4 + 0.07886 * (A - 21)
-    else:
-        beta = 0.0
+        return 0.1102 * (A - 8.7)
+    if A >= 21:
+        return 0.5842 * (A - 21) ** 0.4 + 0.07886 * (A - 21)
+    return 0.0
 
-    return beta
+
+def _kaiser_window_fractional(
+    length: int, side_lobe_suppression_db: float, fractional_delay: float
+) -> np.ndarray:
+    """Create a kaiser window with a fractional offset.
+
+    Parameters
+    ----------
+    length : int
+        Window length.
+    side_lobe_suppression_db : float
+        Expected side lobe suppression in dB.
+    fractional_delay : float
+        Decimal sample offset.
+
+    Returns
+    -------
+    `np.ndarray`
+        Kaiser window.
+
+    """
+    filter_order = length - 1
+    alpha = filter_order / 2
+    beta = _kaiser_window_beta(np.abs(side_lobe_suppression_db))
+    L = np.arange(length).astype(float) - fractional_delay
+
+    if filter_order % 2:
+        L += 0.5
+    else:
+        if fractional_delay > 0.5:
+            L += 1
+    Z = beta * np.sqrt(
+        np.array(1 - ((L - alpha) / alpha) ** 2, dtype="complex")
+    )
+    return np.real(bessel_first_mod(0, Z)) / bessel_first_mod(0, beta)
 
 
 def _indices_above_threshold_dbfs(
@@ -1021,7 +1074,7 @@ def _get_window_envelope(
 def _fractional_delay_filter(
     delay_samples: float,
     filter_order: int,
-    side_lobe_suppression_db: float | None,
+    side_lobe_suppression_db: float,
 ) -> tuple[int, np.ndarray]:
     """This function delivers fractional delay filters according to
     specifications. Besides, additional integer delay, that might be necessary
@@ -1037,7 +1090,7 @@ def _fractional_delay_filter(
     filter_order : int
         Order for the sinc-filter. Higher orders deliver better results but
         require more computational resources.
-    side_lobe_suppression_db : float, optional
+    side_lobe_suppression_db : float
         A kaiser window can be applied to the sinc-filter. Its beta parameter
         will be computed according to the required side lobe suppression (
         a common value would be 60 dB). Pass `None` to avoid any windowing
@@ -1073,18 +1126,9 @@ def _fractional_delay_filter(
     sinc = np.sinc(n)
 
     # =========== Kaiser window ===============================================
-    beta = _kaiser_window_beta(np.abs(side_lobe_suppression_db))
-    alpha = filter_order / 2
-    L = np.arange(filter_order + 1).astype(float) - delay_frac
-    if filter_order % 2:
-        L += 0.5
-    else:
-        if delay_frac > 0.5:
-            L += 1
-    Z = beta * np.sqrt(
-        np.array(1 - ((L - alpha) / alpha) ** 2, dtype="complex")
+    kaiser = _kaiser_window_fractional(
+        filter_order + 1, side_lobe_suppression_db, delay_frac
     )
-    kaiser = np.real(bessel_first_mod(0, Z)) / bessel_first_mod(0, beta)
 
     # Compute filter and final integer delay
     frac_delay_filter = sinc * kaiser