+++ dsptoolbox/classes/filter_class.py
@@ -1,6 +1,7 @@
 """
 Contains Filter class
 """
+
 from pickle import dump, HIGHEST_PROTOCOL
 from warnings import warn
 from copy import deepcopy
@@ -23,7 +24,7 @@ from ._filter import (
 )
 from ._plots import _zp_plot
 from ..plots import general_plot
-from .._general_helpers import _check_format_in_path
+from .._general_helpers import _check_format_in_path, _pad_trim
 
 
 class Filter:
@@ -342,7 +343,7 @@ class Filter:
             polyphase = True
         elif self.filter_type in ("iir", "biquad"):
             if not hasattr(self, "ba"):
-                self.ba = sig.sos2tf(self.sos)
+                self.ba: list = list(sig.sos2tf(self.sos))
             polyphase = False
         else:
             raise ValueError("Wrong filter type for filtering and resampling")
@@ -469,7 +470,7 @@ class Filter:
             self._check_and_update_filter_type()
 
         # Update Metadata about the Filter
-        self.info = filter_configuration
+        self.info: dict = filter_configuration
         self.info["sampling_rate_hz"] = self.sampling_rate_hz
         self.info["filter_type"] = self.filter_type
         if hasattr(self, "ba"):
@@ -542,6 +543,21 @@ class Filter:
             Impulse response of the filter.
 
         """
+        # FIR with no zero phase filtering
+        if self.filter_type == "fir" and not zero_phase:
+            b = self.ba[0].copy()
+            if length_samples < len(b):
+                warn(
+                    f"{length_samples} is not enough for filter with "
+                    + f"length {len(b)}. IR will have the latter length."
+                )
+                length_samples = len(b)
+            b = _pad_trim(b, length_samples)
+            return Signal(
+                None, b, self.sampling_rate_hz, "ir", constrain_amplitude=False
+            )
+
+        # IIR or zero phase IR
         ir_filt = _impulse(length_samples)
         ir_filt = Signal(
             None,
@@ -554,7 +570,7 @@ class Filter:
 
     def get_coefficients(
         self, mode: str = "sos"
-    ) -> list | np.ndarray | tuple[np.ndarray, np.ndarray, np.ndarray]:
+    ) -> list | np.ndarray | tuple[np.ndarray, np.ndarray, np.ndarray] | None:
         """Returns the filter coefficients.
 
         Parameters
@@ -570,6 +586,8 @@ class Filter:
             - `'ba'`: list(b, a) with b and a of type `np.ndarray`.
             - `'sos'`: `np.ndarray`
             - `'zpk'`: tuple(z, p, k) with z, p, k of type `np.ndarray`
+            - Return `None` if user decides that ba->sos is too costly. The
+              threshold is for filters with order > 500.
 
         """
         if mode == "sos":
@@ -798,8 +816,11 @@ class Filter:
             )
         return fig, ax
 
-    def plot_zp(self, show_info_box: bool = False) -> tuple[Figure, Axes]:
-        """Plots zeros and poles with the unit circle.
+    def plot_zp(
+        self, show_info_box: bool = False
+    ) -> tuple[Figure, Axes] | None:
+        """Plots zeros and poles with the unit circle. This returns `None` and
+        produces no plot if user decides that conversion ba->sos is too costly.
 
         Parameters
         ----------